Move common preference screen logic to BaseSettingsFragment
