+++ app/src/main/java/com/twilio/video/app/ui/settings/BandwidthProfileSettingsFragment.kt
@@ -1,9 +1,7 @@
 package com.twilio.video.app.ui.settings
 
 import android.os.Bundle
-import android.view.MenuItem
 import com.twilio.video.app.R
-import com.twilio.video.app.base.BaseActivity
 import com.twilio.video.app.data.Preferences
 
 class BandwidthProfileSettingsFragment : BaseSettingsFragment() {
@@ -34,19 +32,5 @@ class BandwidthProfileSettingsFragment : BaseSettingsFragment() {
         setListPreferenceValue(R.array.settings_screen_bandwidth_profile_render_dimensions,
             Preferences.BANDWIDTH_PROFILE_HIGH_TRACK_PRIORITY_RENDER_DIMENSIONS,
             Preferences.BANDWIDTH_PROFILE_HIGH_TRACK_PRIORITY_RENDER_DIMENSIONS_DEFAULT)
-    }
-
-    override fun onResume() {
-        super.onResume()
-
-        (requireActivity() as BaseActivity).supportActionBar?.title = preferenceScreen.title
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == android.R.id.home) {
-            parentFragmentManager.popBackStack()
-            return true
-        }
-        return false
     }
 }
